requirements.txt
Flask==3.0.2
langchain==0.1.10
langchain-openai==0.0.8
python-dotenv==1.0.1
chromadb==0.4.24
requests==2.31.0
